gtk fixes: in Makefile - fix datadir for mingw users - remove the shared lib during the clean rule - use $(ENCODE_BIN) instead of x264_gtk_encode - add some $(DESTDIR) and create some directories when necessary - remove -lintl statfile_length -> statsfile_length fix the "sensitivity" of the widget of update_statfile the logo is now handled correctly on windows added: beginning of multipass support patch by Vincent Torri.
git-svn-id: svn://svn.videolan.org/x264/trunk@580 df754926-b1dd-0310-bc7b-ec298dee348c

+++ gtk/x264_gtk_encode_encode.c
@@ -140,7 +140,7 @@ x264_gtk_encode_encode (X264_Thread_Data *thread_data)
                                                  sizeof (X264_Pipe_Data),
                                                  &size, NULL);
               if (status != G_IO_STATUS_NORMAL) {
-                g_print (_("Error ! %d %d %d\n"), status, sizeof (X264_Pipe_Data), size);
+                g_print (_("Error ! %d %d %d\n"), status, (int)sizeof (X264_Pipe_Data), (int)size);
               }
               else {
                 /* we force the GIOChannel to write to the pipeline */
